A step on the tuning ui
1. Tab layout
2. Show the tuning grid
3. start cleaning up the code but still a mess

Addresses #4390
